added start_game

+++ src/schema.rs
@@ -2,7 +2,7 @@
 
 diesel::table! {
     games (channel_id) {
-        channel_id -> Integer,
+        channel_id -> BigInt,
         players -> Nullable<Text>,
         board -> Nullable<Text>,
         game -> Nullable<Text>,
